+++ restful-test-instances/organisations/src/main/java/org/restful/test/instances/service/OrganizationService.java
@@ -31,7 +31,7 @@ import static org.hibernate.validator.internal.util.Contracts.assertNotNull;
         makeFinal = true
 )
 @Service
-public final class OrganizationService {
+public class OrganizationService {
 
     OrganizationRepository organizationRepository;