+++ restful-test-instances/service-test/src/test/java/org/restful/test/instances/service/test/service/OrganizationServiceTest.java
@@ -17,14 +17,17 @@ import org.restful.test.instances.service.mapping.OrganizationModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.data.jpa.domain.Specification;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.junit4.SpringRunner;
 
+import java.util.Collections;
+import java.util.List;
 import java.util.Optional;
 
 import static org.hibernate.validator.internal.util.Contracts.assertNotNull;
 import static org.hibernate.validator.internal.util.Contracts.assertTrue;
-import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
@@ -83,6 +86,18 @@ public class OrganizationServiceTest {
                 );
         when(organizationRepository.findById(Long.valueOf(2)))
                 .thenReturn(Optional.ofNullable(null));
+        when(organizationRepository.findAll((Specification<Organization>) null))
+                .thenReturn(Collections.singletonList(
+                        Organization.builder()
+                                .uid(Long.valueOf(1))
+                                .build()
+                ));
+        when(organizationRepository.findAll(any(Specification.class)))
+                .thenReturn(Collections.singletonList(
+                        Organization.builder()
+                                .uid(Long.valueOf(1))
+                                .build()
+                ));
         when(organizationModelMapper.map(any(OrganizationDetail.class), any(Organization.class)))
                 .thenReturn(Organization.builder()
                         .uid(Long.valueOf(1))
@@ -93,6 +108,12 @@ public class OrganizationServiceTest {
                         .uid(Optional.ofNullable(Long.valueOf(1)))
                         .build()
                 );
+        when(organizationModelMapper.map(anyCollection(), eq(OrganizationDetail.class)))
+                .thenReturn(Collections.singletonList(
+                        OrganizationDetail.builder()
+                                .uid(Optional.ofNullable(Long.valueOf(1)))
+                                .build()
+                ));
     }
 
     /**
@@ -210,5 +231,61 @@ public class OrganizationServiceTest {
         );
 
         verify(organizationRepository).findById(uidOrganization);
+    }
+
+    /**
+     * Find positive when all request params is null then successful find all.
+     *
+     * @throws CustomModelMapper.MappingException the mapping exception
+     */
+    @Test
+    public void find_positive_whenAllRequestParamsIsNull_thenSuccessfulFound()
+            throws CustomModelMapper.MappingException {
+        List<OrganizationDetail> organizationResponse = this.organizationService.find(
+                null,
+                null,
+                null,
+                null,
+                null
+        );
+        verify(organizationRepository).findAll((Specification<Organization>) null);
+        verify(organizationModelMapper).map(anyCollection(), eq(OrganizationDetail.class));
+        assertNotNull(organizationResponse, "Что-то пошло не так");
+        assertTrue(!organizationResponse.isEmpty(), "В ответе ничего нет");
+    }
+
+    /**
+     * Find positive when request params is not null then successful find all by params.
+     *
+     * @throws CustomModelMapper.MappingException the mapping exception
+     */
+    @Test
+    public void find_positive_whenRequestParamsIsNotNull_thenSuccessfulFindAllByParams()
+            throws CustomModelMapper.MappingException {
+        List<Long> uid = Collections.singletonList(1L);
+        List<String> name = Collections.singletonList("WCorp");
+        List<String> inn = Collections.singletonList("01");
+        List<String> kpp = Collections.singletonList("01");
+        List<String> address = Collections.singletonList(".ell road");
+        log.info(
+                "Создали параметры для запроса:\nuids - {}\nname - {}\ninn - {}\nkpp - {}\naddress - {}",
+                uid,
+                name,
+                inn,
+                kpp,
+                address
+                );
+
+        List<OrganizationDetail> organizationResponse = this.organizationService.find(
+                uid,
+                name,
+                inn,
+                kpp,
+                address
+        );
+        verify(organizationRepository).findAll(any(Specification.class));
+        verify(organizationModelMapper).map(anyCollection(), eq(OrganizationDetail.class));
+        assertNotNull(organizationResponse, "Что-то пошло не так");
+        assertTrue(!organizationResponse.isEmpty(), "В ответе нчиего нет");
     }
 }