+++ restful-test-instances/CHANGELOG.md
@@ -10,6 +10,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 - The tests for other organisation service methods.
 
+## [0.7.0] - 2020-09-05
+
+### Added
+
+- The uint tests of CRUD organisation controller methods for the organization creating, updating, deleting or finding;
+- The integration tests of organisation controller layer.
+
 ## [0.6.0] - 2020-09-05
 
 ### Changed